chore : contest 2170b had to change the type int to long long to avoid overflow

+++ round_185(div2)/contest_2170b.cpp
@@ -1,17 +1,15 @@
-#include <iostream>
-#include <vector>
-#include <algorithm>
+#include <bits/stdc++.h>
 using namespace std;
 
 int main() {
-    int t;
+    long  t;
     cin >> t;
     while (t--) {
-        int n;
+        long n;
         cin >> n;
         vector<int> b(n);
-        int zeroes = 0;
-        int sum = 0;
+        long long zeroes = 0;
+        long long sum = 0;
         for (int i = 0; i < n; i++) {
             cin >> b[i];
             if(b[i] == 0) zeroes++;
